Added Mentalist
